requirements.txt
discord.py
python-dotenv
yt-dlp
spotipy
PyNaCl
opuslib

